# swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "A web service that exposes Vehicle (Bus), Fleet (Operator) and Activity (Stop) data, for a given time frame",
    "version" : "1.0.0",
    "title" : "Bus GPS API"
  },
  "basePath" : "/api",
  "schemes" : [ "http", "https" ],
  "consumes" : [ "application/json" ],
  "produces" : [ "application/json" ],
  "paths" : {
    "/operators" : {
      "get" : {
        "tags" : [ "Operators" ],
        "summary" : "Returns a list of running operators.",
        "description" : "Returns a list of running operators.",
        "operationId" : "operators",
        "parameters" : [ {
          "name" : "start-time",
          "in" : "header",
          "description" : "Start time (2012-12-31).",
          "required" : true,
          "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
          "type" : "string"
        }, {
          "name" : "end-time",
          "in" : "header",
          "description" : "End time (2013-01-31).",
          "required" : true,
          "type" : "string"
        }],
        "responses" : {
          "200" : {
            "description" : "A list of operators.",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/OperatorResponse"
              }
            }
          },
          "400" : {
            "description" : "Validation error"
          },
          "500" : {
            "description" : "Internal server error"
          }
        }
      }
    },
    "/vehicles" : {
      "get" : {
        "tags" : [ "Vehicles" ],
        "summary" : "Returns a list of vehicles",
        "description" : "Returns a list of vehicles",
        "operationId" : "vehicles",
        "parameters" : [ {
          "name" : "start-time",
          "in" : "header",
          "description" : "Start time (2012-12-31).",
          "required" : true,
          "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
          "type" : "string"
        }, {
          "name" : "end-time",
          "in" : "header",
          "description" : "End time (2013-01-31).",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "operator",
          "in" : "header",
          "description" : "An operator (RD)",
          "required" : true,
          "type" : "string"
        }],
        "responses" : {
          "200" : {
            "description" : "A list of vehicles ids.",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/VehicleResponse"
              }
            }
          },
          "400" : {
            "description" : "Validation error"
          },
          "500" : {
            "description" : "Internal server error"
          }
        }
      }
    },
    "/vehiclesAtStop" : {
      "get" : {
        "tags" : [ "Vehicles at a Stop" ],
        "summary" : "Return which vehicles are at a stop.",
        "description" : "Return which vehicles are at a stop.",
        "operationId" : "vehicleAtStop",
        "parameters" : [ {
          "name" : "start-time",
          "in" : "header",
          "description" : "Start time (2012-12-31).",
          "required" : true,
          "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
          "type" : "string"
        }, {
          "name" : "end-time",
          "in" : "header",
          "description" : "End time (2013-01-31).",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "fleet",
          "in" : "header",
          "description" : "A fleet (RD)",
          "required" : true,
          "type" : "string"
        }],
        "responses" : {
          "200" : {
            "description" : "A list of vehicles at a Stop.",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/VehicleStopResponse"
              }
            }
          },
          "400" : {
            "description" : "Validation error"
          },
          "500" : {
            "description" : "Internal server error"
          }
        }
      }
    },
    "/vehicleTrace" : {
      "get" : {
        "tags" : [ "Trace of a vehicle" ],
        "summary" : "Return the trace of that vehicle",
        "description" : "Return the trace of that vehicle",
        "operationId" : "vehicleTrace",
        "parameters" : [ {
          "name" : "start-time",
          "in" : "header",
          "description" : "Start time (2012-12-31).",
          "required" : true,
          "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
          "type" : "string"
        }, {
          "name" : "end-time",
          "in" : "header",
          "description" : "End time (2013-01-31).",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "vehicleId",
          "in" : "header",
          "description" : "A vehicle id (123445)",
          "required" : true,
          "type" : "string"
        }],
        "responses" : {
          "200" : {
            "description" : "A list of GPS entries of that vehicle.",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/VehiclePositionResponse"
              }
            }
          },
          "400" : {
            "description" : "Validation error"
          },
          "500" : {
            "description" : "Internal server error"
          }
        }
      }
    }
  },
  "definitions" : {
    "OperatorResponse" : {
      "type" : "object",
      "properties" : {
        "operator" : {
          "type" : "string",
          "example" : "RD"
        }
      }
    },
    "VehicleResponse" : {
      "type" : "object",
      "properties" : {
        "vehicleId" : {
          "type" : "string",
          "example" : "1250"
        }
      }
    },
    "VehicleStopResponse" : {
      "type" : "object",
      "properties" : {
        "vehicleId" : {
          "type" : "string",
          "example" : "12233"
        },
        "atStop" : {
          "type" : "boolean",
          "example" : true
        }
      }
    },
    "VehiclePositionResponse" : {
      "type" : "object",
      "properties" : {
        "longitude" : {
          "type" : "string",
          "example" : "-6.257359"
        },
        "latitude" : {
          "type" : "string",
          "example" : "53.373501"
        }
      }
    }
  }
}